+++ logprep/connector/confluent_kafka/input.py
@@ -130,11 +130,16 @@ class ConfluentKafkaInput(Input):
 
         librdkafka_replyq: GaugeMetric = field(
             factory=lambda: GaugeMetric(
-                description="Number of ops (callbacks, events, etc) waiting in queue for application to serve with rd_kafka_poll()",
+                description=(
+                    "Number of ops (callbacks, events, etc) waiting in "
+                    "queue for application to serve with rd_kafka_poll()"
+                ),
                 name="confluent_kafka_input_librdkafka_replyq",
             )
         )
-        """Number of ops (callbacks, events, etc) waiting in queue for application to serve with rd_kafka_poll()"""
+        """Number of ops (callbacks, events, etc) waiting in queue for application
+           to serve with rd_kafka_poll()
+        """
         librdkafka_tx: GaugeMetric = field(
             factory=lambda: GaugeMetric(
                 description="Total number of requests sent to Kafka brokers",
@@ -166,14 +171,22 @@ class ConfluentKafkaInput(Input):
         """Total number of bytes received from Kafka brokers"""
         librdkafka_rxmsgs: GaugeMetric = field(
             factory=lambda: GaugeMetric(
-                description="Total number of messages consumed, not including ignored messages (due to offset, etc), from Kafka brokers.",
+                description=(
+                    "Total number of messages consumed, not including ignored messages"
+                    "(due to offset, etc), from Kafka brokers."
+                ),
                 name="confluent_kafka_input_librdkafka_rxmsgs",
             )
         )
-        """Total number of messages consumed, not including ignored messages (due to offset, etc), from Kafka brokers."""
+        """Total number of messages consumed, not including ignored messages
+           (due to offset, etc), from Kafka brokers.
+        """
         librdkafka_rxmsg_bytes: GaugeMetric = field(
             factory=lambda: GaugeMetric(
-                description="Total number of message bytes (including framing) received from Kafka brokers",
+                description=(
+                    "Total number of message bytes (including framing)"
+                    "received from Kafka brokers"
+                ),
                 name="confluent_kafka_input_librdkafka_rxmsg_bytes",
             )
         )
@@ -195,11 +208,11 @@ class ConfluentKafkaInput(Input):
         """Time elapsed since last rebalance (assign or revoke) (milliseconds)."""
         librdkafka_cgrp_rebalance_cnt: GaugeMetric = field(
             factory=lambda: GaugeMetric(
-                description="Total number of rebalances (assign or revoke).",
+                description="Total number of rebalance (assign or revoke).",
                 name="confluent_kafka_input_librdkafka_cgrp_rebalance_cnt",
             )
         )
-        """Total number of rebalances (assign or revoke)."""
+        """Total number of rebalance (assign or revoke)."""
         librdkafka_cgrp_assignment_size: GaugeMetric = field(
             factory=lambda: GaugeMetric(
                 description="Current assignment's partition count.",
@@ -316,7 +329,7 @@ class ConfluentKafkaInput(Input):
             the error that occurred
         """
         self.metrics.number_of_errors += 1
-        logger.error(f"{self.describe()}: {error}")
+        logger.error("%s: %s", self.describe(), error)
 
     def _stats_callback(self, stats_raw: str) -> None:
         """Callback for statistics data. This callback is triggered by poll()
@@ -487,12 +500,13 @@ class ConfluentKafkaInput(Input):
         except KafkaException as error:
             raise InputWarning(self, f"{error}, {self._last_valid_record}") from error
 
-    def _assign_callback(self, consumer, topic_partitions):
+    def _assign_callback(self, topic_partitions: list[TopicPartition]) -> None:
         for topic_partition in topic_partitions:
             offset, partition = topic_partition.offset, topic_partition.partition
+            member_id = self._get_memberid()
             logger.info(
                 "%s was assigned to topic: %s | partition %s",
-                consumer.memberid(),
+                member_id,
                 topic_partition.topic,
                 partition,
             )
@@ -503,26 +517,37 @@ class ConfluentKafkaInput(Input):
             self.metrics.committed_offsets.add_with_labels(offset, labels)
             self.metrics.current_offsets.add_with_labels(offset, labels)
 
-    def _revoke_callback(self, consumer, topic_partitions):
+    def _revoke_callback(self, topic_partitions: list[TopicPartition]) -> None:
+
         for topic_partition in topic_partitions:
             self.metrics.number_of_warnings += 1
+            member_id = self._get_memberid()
             logger.warning(
                 "%s to be revoked from topic: %s | partition %s",
-                consumer.memberid(),
+                member_id,
                 topic_partition.topic,
                 topic_partition.partition,
             )
         self.batch_finished_callback()
 
-    def _lost_callback(self, consumer, topic_partitions):
+    def _lost_callback(self, topic_partitions: list[TopicPartition]) -> None:
         for topic_partition in topic_partitions:
             self.metrics.number_of_warnings += 1
+            member_id = self._get_memberid()
             logger.warning(
                 "%s has lost topic: %s | partition %s - try to reassign",
-                consumer.memberid(),
+                member_id,
                 topic_partition.topic,
                 topic_partition.partition,
             )
+
+    def _get_memberid(self) -> str | None:
+        member_id = None
+        try:
+            member_id = self._consumer.memberid()
+        except RuntimeError as error:
+            logger.error("Failed to retrieve member ID: %s", error)
+        return member_id
 
     def shut_down(self) -> None:
         """Close consumer, which also commits kafka offsets."""