+++ tests/unit/connector/test_confluent_kafka_common.py
@@ -34,7 +34,7 @@ class CommonConfluentKafkaTestCase:
             test_error = Exception("test error")
             self.object._error_callback(test_error)
             mock_error.assert_called()
-            mock_error.assert_called_with(f"{self.object.describe()}: {test_error}")
+            mock_error.assert_called_with("%s: %s", self.object.describe(), test_error)
         assert self.object.metrics.number_of_errors == 1
 
     def test_stats_callback_sets_metric_objetc_attributes(self):