+++ tests/unit/connector/test_confluent_kafka_input.py
@@ -4,12 +4,13 @@
 # pylint: disable=wrong-import-order
 # pylint: disable=attribute-defined-outside-init
 import os
+import re
 import socket
 from copy import deepcopy
 from unittest import mock
 
 import pytest
-from confluent_kafka import OFFSET_BEGINNING, KafkaError, KafkaException
+from confluent_kafka import OFFSET_BEGINNING, KafkaError, KafkaException  # type: ignore
 
 from logprep.abc.input import (
     CriticalInputError,
@@ -345,7 +346,7 @@ class TestConfluentKafkaInput(BaseInputTestCase, CommonConfluentKafkaTestCase):
         self.object.metrics.number_of_warnings = 0
         mock_partitions = [mock.MagicMock()]
         with mock.patch("logging.Logger.warning") as mock_warning:
-            self.object._lost_callback(mock_consumer, mock_partitions)
+            self.object._lost_callback(mock_partitions)
         mock_warning.assert_called()
         assert self.object.metrics.number_of_warnings == 1
 
@@ -360,14 +361,13 @@ class TestConfluentKafkaInput(BaseInputTestCase, CommonConfluentKafkaTestCase):
         }
         self.object.metrics.committed_offsets.add_with_labels.assert_called_with(0, expected_labels)
 
-    @mock.patch("logprep.connector.confluent_kafka.input.Consumer")
-    def test_assign_callback_sets_offsets_and_logs_info(self, mock_consumer):
+    def test_assign_callback_sets_offsets_and_logs_info(self):
         self.object.metrics.committed_offsets.add_with_labels = mock.MagicMock()
         self.object.metrics.current_offsets.add_with_labels = mock.MagicMock()
         mock_partitions = [mock.MagicMock()]
         mock_partitions[0].offset = OFFSET_BEGINNING
         with mock.patch("logging.Logger.info") as mock_info:
-            self.object._assign_callback(mock_consumer, mock_partitions)
+            self.object._assign_callback(mock_partitions)
         expected_labels = {
             "description": f"topic: test_input_raw - partition: {mock_partitions[0].partition}"
         }
@@ -375,23 +375,29 @@ class TestConfluentKafkaInput(BaseInputTestCase, CommonConfluentKafkaTestCase):
         self.object.metrics.committed_offsets.add_with_labels.assert_called_with(0, expected_labels)
         self.object.metrics.current_offsets.add_with_labels.assert_called_with(0, expected_labels)
 
-    @mock.patch("logprep.connector.confluent_kafka.input.Consumer")
-    def test_revoke_callback_logs_warning_and_counts(self, mock_consumer):
+    def test_revoke_callback_logs_warning_and_counts(self):
         self.object.metrics.number_of_warnings = 0
         self.object.output_connector = mock.MagicMock()
         mock_partitions = [mock.MagicMock()]
         with mock.patch("logging.Logger.warning") as mock_warning:
-            self.object._revoke_callback(mock_consumer, mock_partitions)
+            self.object._revoke_callback(mock_partitions)
         mock_warning.assert_called()
         assert self.object.metrics.number_of_warnings == 1
 
-    @mock.patch("logprep.connector.confluent_kafka.input.Consumer")
-    def test_revoke_callback_calls_batch_finished_callback(self, mock_consumer):
+    def test_revoke_callback_calls_batch_finished_callback(self):
         self.object.output_connector = mock.MagicMock()
         self.object.batch_finished_callback = mock.MagicMock()
         mock_partitions = [mock.MagicMock()]
-        self.object._revoke_callback(mock_consumer, mock_partitions)
+        self.object._revoke_callback(mock_partitions)
         self.object.batch_finished_callback.assert_called()
+
+    def test_revoke_callback_logs_error_if_consumer_closed(self, caplog):
+        with mock.patch.object(self.object, "_consumer") as mock_consumer:
+            mock_consumer.memberid = mock.MagicMock()
+            mock_consumer.memberid.side_effect = RuntimeError("Consumer is closed")
+            mock_partitions = [mock.MagicMock()]
+            self.object._revoke_callback(mock_partitions)
+            assert re.search(r"ERROR.*Consumer is closed", caplog.text)
 
     def test_health_returns_true_if_no_error(self):
         self.object._admin = mock.MagicMock()