+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Improvements
 ### Bugfix
 
+- Fixed logging error in _revoke_callback() by adding error handling
+
 ## 16.1.0
 ### Deprecations